requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.0.0
appnope==0.1.3
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
decorator==4.4.2
deprecation==2.1.0
elevenlabs==0.2.26
executing==2.0.0
Flask==3.0.0
Flask-Cors==4.0.0
fonttools==4.43.1
frozenlist==1.4.0
gotrue==1.2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
imageio==2.31.6
imageio-ffmpeg==0.4.9
ipython==8.16.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
moviepy==1.0.3
multidict==6.0.4
nltk==3.8.1
numpy==1.26.1
openai==0.28.1
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
plotly==5.17.0
postgrest==0.11.0
proglog==0.1.10
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
pytz==2023.3.post1
realtime==1.0.0
regex==2023.10.3
replicate==0.15.4
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
storage3==0.6.1
StrEnum==0.4.15
supabase==1.2.0
supafunc==0.2.3
tenacity==8.2.3
threadpoolctl==3.2.0
tqdm==4.66.1
traitlets==5.11.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
wcwidth==0.2.8
websockets==12.0
Werkzeug==3.0.0
yarl==1.9.2
